modules/strands-mcp-inter-agent/pyproject.toml
[project]
name = "strands-mcp-inter-agent"
version = "0.0.1"
requires-python = ">=3.13"
dependencies = ["mcp", "strands-agents"]

[dependency-groups]
debug = ["bedrock-agentcore"]
deploy = ["aws-cdk-lib", "constructs", "cdk-ecr-deployment"]
